# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Saukele Digital Registry API
  version: 1.0.0
  description: |
    REST API for **Saukele** — a digital wedding registry that models Kazakh
    family-tree kinship, multi-currency pool funding with snapshotted exchange
    rates, and traditional-gift logistics.

    ## Conventions
    - All timestamps are ISO-8601 UTC.
    - All amounts are decimals serialized as JSON numbers; the canonical currency
      is KZT. Foreign currencies are stored alongside the rate that was active at
      transaction time and are NEVER recalculated.
    - All endpoints (except `/auth/register`, `/auth/login`, `/auth/refresh`,
      `/healthz`) require a valid JWT access token.
    - Pagination on list endpoints is **cursor-based** (keyset). The cursor is
      opaque to the client — pass back exactly what `next_cursor` returned.
    - Errors follow a single shape (`ErrorResponse`).

  contact:
    name: Saukele Backend Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.saukele.kz/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CursorParam:
      name: cursor
      in: query
      required: false
      description: Opaque keyset cursor returned by the previous page response.
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    GiftIdPath:
      name: gift_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, code, status_code]
      properties:
        error:
          type: string
          description: Human-readable summary.
          example: "Validation failed"
        code:
          type: string
          description: Stable machine-readable code (UPPER_SNAKE).
          example: "VALIDATION_ERROR"
        status_code:
          type: integer
          example: 422
        details:
          type: array
          description: Field-level errors when applicable.
          items:
            type: object
            properties:
              field: { type: string, example: "email" }
              issue: { type: string, example: "must be a valid email" }
        request_id:
          type: string
          example: "req_01HXYZ..."
    Pagination:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: {}
        next_cursor:
          type: string
          nullable: true
          description: Pass to `?cursor=` for the next page. `null` = last page.
...
    GiftState:
      type: string
      enum: [PENDING, FUNDED, PURCHASED, DELIVERED, CANCELLED]
    Gift:
      type: object
      properties:
        id:                  { type: string, format: uuid }
        registry_id:         { type: string, format: uuid }
        title:               { type: string, example: "Сәукеле (traditional headpiece)" }
        description:         { type: string, nullable: true }
        target_amount_kzt:   { type: number, example: 450000 }
        current_amount_kzt:  { type: number, example: 120000 }
        required_tier_rank:  { type: integer, example: 1 }
        is_fragile:          { type: boolean }
        state:               { $ref: '#/components/schemas/GiftState' }
        created_at:          { type: string, format: date-time }
...
    ContributionStatus:
      type: string
      enum: [PENDING, FUNDED, FAILED, REFUNDED]
    Contribution:
      type: object
      properties:
        id:                  { type: string, format: uuid }
        gift_id:             { type: string, format: uuid }
        guest_id:            { type: string, format: uuid }
        amount_kzt:          { type: number, example: 45000.00 }
        amount_original:     { type: number, example: 100.00 }
        currency_original:   { type: string, example: "EUR" }
        exchange_rate:       { type: number, example: 450.000000 }
        rate_locked_at:      { type: string, format: date-time }
        status:              { $ref: '#/components/schemas/ContributionStatus' }
        created_at:          { type: string, format: date-time }
    ContributionCreate:
      type: object
      required: [guest_id, amount_original, currency_original]
      properties:
        guest_id:          { type: string, format: uuid }
        amount_original:   { type: number, minimum: 0.01 }
        currency_original:
          type: string
          minLength: 3
          maxLength: 3
          example: "EUR"
          description: ISO 4217 code. Must be supported by the exchange provider.

  # ── Reusable error responses ────────────────────────────────────────────────
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid access token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Authenticated, but role or tier does not permit this action.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource does not exist.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: State conflict (e.g., gift already funded, race condition).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unprocessable:
      description: Validation passed type-wise, but business rule rejected the request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
security:
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /gifts/{gift_id}:
    parameters:
      - $ref: '#/components/parameters/GiftIdPath'
    get:
      tags: [Gifts]
      summary: Get a single gift
      responses:
        '200':
          description: Gift details.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Gift' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Contributions (the core business transaction) ─────────────────────────
  /gifts/{gift_id}/contributions:
    parameters:
      - $ref: '#/components/parameters/GiftIdPath'

    get:
      tags: [Contributions]
      summary: List contributions for a gift (cursor-paginated)
      description: |
        Returns contributions to a pooled gift using cursor pagination keyed on
        `created_at` for deterministic ordering.
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ContributionStatus' }
      responses:
        '200':
          description: Page of contributions.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Pagination'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Contribution' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }

    post:
      tags: [Contributions]
      summary: Contribute to a gift pool (atomic, multi-currency snapshot)
      description: |
        Runs inside a Serializable Postgres transaction with `SELECT … FOR UPDATE`.
        The exchange rate is fetched and **snapshotted** into the row at write
        time and is never recalculated.

        **Failure modes:**
        - `409 GIFT_NOT_OPEN` — gift is not in PENDING state.
        - `422 POOL_CAP_EXCEEDED` — would overfund the gift.
        - `422 UNSUPPORTED_CURRENCY` — exchange provider does not support the code.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContributionCreate' }
            examples:
              eur:
                value:
                  guest_id: "d290f1ee-6c54-4b01-90e6-d701748f0851"
                  amount_original: 100.00
                  currency_original: "EUR"
      responses:
        '201':
          description: Contribution recorded.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contribution' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ── Guests & Family Tree ──────────────────────────────────────────────────
...